Se agregan validaciones de tipo pago y de hora, se agrega el campo cantidadAutorizada en la tabla Evento, Falta validar el evento
