+++ app/src/main/res/layout/activity_evento_eliminar.xml
@@ -9,6 +9,16 @@
     tools:context=".Chris.Evento.EventoEliminarActivity">
 
     <TextView
+        android:id="@+id/textView_titulo"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:layout_gravity="center"
+        android:text="@string/titulo_evento_eliminar"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:textColor="#000000" />
+
+    <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/NumEvenTxt"
