+++ app/src/main/res/layout/activity_evento_insertar.xml
@@ -9,6 +9,16 @@
     tools:context=".Chris.Evento.EventoInsertarActivity">
 
     <TextView
+        android:id="@+id/textView_titulo"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:layout_gravity="center"
+        android:text="@string/titulo_evento_insertar"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:textColor="#000000" />
+
+    <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/NumEvenTxt"
@@ -60,6 +70,19 @@
         android:hint="@string/CostoEventoHint"
         android:inputType="numberDecimal"
         android:maxLength="4"/>
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/cantAutEventTxt"
+        android:textColor="@color/black"/>
+
+    <EditText
+        android:id="@+id/EditCantAutorEvento"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:hint="@string/Hint_cantAutorizada"
+        android:inputType="number"/>
 
     <Button
         android:id="@+id/botonAgregarNuevoEvento"
