+++ app/src/main/res/layout/activity_tipo_pago_eliminar.xml
@@ -9,6 +9,17 @@
     tools:context=".Chris.TipoPago.TipoPagoEliminarActivity">
 
     <TextView
+        android:id="@+id/textView_titulo"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:layout_gravity="center"
+        android:text="@string/titulo_tipoPago_eliminar"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:textColor="#000000" />
+
+    <TextView
+        android:paddingTop="20dp"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/TipoPagoIdentificador"
